+++ Gateway/Request/TransactionDataBuilder.php
@@ -83,7 +83,7 @@ class TransactionDataBuilder implements BuilderInterface
 
         $amount = $this->subjectReader->readAmount($buildSubject);
         if ($amount) {
-            $transactionData[self::$transaction][self::$amount] = $this->formatPrice($amount) * 100;
+            $transactionData[self::$transaction][self::$amount] = (int)($this->formatPrice($amount) * 100);
         }
 
         return $transactionData;
